+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/key_client.hpp
@@ -74,6 +74,13 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
         KeyClientOptions options = KeyClientOptions());
 
     /**
+     * @brief Construct a new Key Client object from another key client.
+     *
+     * @param keyClient An existing key vault key client.
+     */
+    explicit KeyClient(KeyClient const& keyClient) : m_pipeline(keyClient.m_pipeline) {}
+
+    /**
      * @brief Gets the public part of a stored key.
      *
      * @remark The get key operation is applicable to all key types. If the requested key is
@@ -168,15 +175,14 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @remark Use \p options to control which page to get. If
      * #GetPropertiesOfKeysSinglePageOptions.NextLink is not set, the operation will get the first
-     * page and it will set the `ContinuationToken` from the #KeyPropertiesSinglePage as the next
+     * page and it will set the `NextPageToken` from the #KeyPropertiesSinglePage as the next
      * page of the response if there is a next page.
      *
      * @param options The #GetPropertiesOfKeysSinglePageOptions object to for setting the operation
      * up.
      * @param context A #Azure::Core::Context controlling the request lifetime.
-     * @return Azure::Response<KeyPropertiesSinglePage>
      */
-    Azure::Response<KeyPropertiesSinglePage> GetPropertiesOfKeysSinglePage(
+    KeyPropertiesSinglePage GetPropertiesOfKeysSinglePage(
         GetPropertiesOfKeysSinglePageOptions const& options
         = GetPropertiesOfKeysSinglePageOptions(),
         Azure::Core::Context const& context = Azure::Core::Context()) const;
@@ -191,16 +197,15 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @remark Use \p options to control which page to get. If
      * #GetPropertiesOfKeyVersionsSinglePageOptions.NextLink is not set, the operation will get the
-     * first page and it will set the `ContinuationToken` from the #KeyPropertiesSinglePage as the
+     * first page and it will set the `NextPageToken` from the #KeyPropertiesSinglePage as the
      * next page of the response if there is a next page.
      *
      * @param name The name of the key.
      * @param options The #GetPropertiesOfKeyVersionsSinglePageOptions object to for setting the
      * operation up.
      * @param context A #Azure::Core::Context controlling the request lifetime.
-     * @return Azure::Response<KeyPropertiesSinglePage>
      */
-    Azure::Response<KeyPropertiesSinglePage> GetPropertiesOfKeyVersionsSinglePage(
+    KeyPropertiesSinglePage GetPropertiesOfKeyVersionsSinglePage(
         std::string const& name,
         GetPropertiesOfKeyVersionsSinglePageOptions const& options
         = GetPropertiesOfKeyVersionsSinglePageOptions(),
@@ -262,15 +267,15 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      * requires the keys/list permission.
      *
      * @remark Use \p options to control which page to get. If
-     * #GetPropertiesOfKeyVersionsSinglePageOptions.NextLink is not set, the operation will get the
-     * first page and it will set the `ContinuationToken` from the #KeyPropertiesSinglePage as the
+     * #GetPropertiesOfKeyVersionsSinglePageOptions.NextPageToken is not set, the operation will get
+     * the first page and it will set the `NextPageToken` from the #KeyPropertiesSinglePage as the
      * next page of the response if there is a next page.
      *
      * @param options The #GetDeletedKeysSinglePageOptions object to for setting the operation up.
      * @param context A #Azure::Core::Context controlling the request lifetime.
      * @return Azure::Response<DeletedKeySinglePage>
      */
-    Azure::Response<DeletedKeySinglePage> GetDeletedKeysSinglePage(
+    DeletedKeySinglePage GetDeletedKeysSinglePage(
         GetDeletedKeysSinglePageOptions const& options = GetDeletedKeysSinglePageOptions(),
         Azure::Core::Context const& context = Azure::Core::Context()) const;